requirements.txt
# FastAPI + ASGI
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Encryption
cryptography==43.0.1

# Task queue
celery[redis]==5.4.0
redis==5.1.0

# HTTP client (for exchange APIs)
httpx==0.27.2
requests==2.32.3

# Settings
pydantic-settings==2.5.2

# Rate limiting
slowapi==0.1.9

# WebSocket
websockets==13.0

# Utilities
python-dotenv==1.0.1
